refactor: move program inputs to the top of core crate
